Bugs fixed, code running
The bugs are all fixed. THe code is not 100% clean, I'm getting warnings. Next step is to correct the log-uniform distribution to output the right values. If it's working properly, there should be an equal chance of getting a number between 0 and 1, or 1 and 10, or 10 and 100. The next step after that is to do a proper simulation test, comparing Helmut's code output to mine.