Got log uniform and Moore-Penrose to work.
The file "Eigen" is a portable lightweight easy to use linear algebra library that does not require linking, all it requires is that the include path of the compiler includes Eigen. It does the Moore Penrose pseudoinverse in a very simple easy to understand format. It is working correctly here. I also got the log uniform to work correctly, the main.cpp includes a test for this (the instances of nums between 0 and 10 equals the number between 10 and 100 and 100 and 1000, approximately). Also, this version includes a test if the Delaunay Triangulation fails to connect a node to another node. We want control over how many springs there is ourselves and might want to eliminate some springs/nodes afterwards, but the simulation should start with every node connected to at least one other node.