The code is now outputting a weight matrix
The code is now finally outputting a weight matrix for the learning phase of reservoir computing. However, sometimes there are bad simulations because the solution explodes. And also some nodes are unconnected, they are connected to the edges of the frame.  The former means that the weight matrix is all 0 so the simulation can be discarded safely. The latter does not effect the output for connected nodes, it's just unrealistic and sort of a waste as nodes are not connected to the edge of anything in real life, the frame does not exist. I think we can work with what we have but it is not optimal, my priority is to get solutions working for Oxford but this problem cannot be neglected.