Oscillations.
Getting oscillatory behaviour in the Fsum, the sum of the forces on one of the springs, so it looks hopeful. Also have weight vector (here called Learning_Weights), so very close to getting proper reservoir computation. It does look like it will explode to infinity for certain values of k3 and d3 as near t=0, these values do not matter but as it goes  towards 1 than x3 term becomes more and more important and dominats, and k3 and d3 are large.