FIguring out why force explodes
The positions of the nodes are fixed for the leftmost and rightmost ones and that fixes the whole mass spring. So it is not that the nodes are going off to infinity, taht is not the problem. The problem is the force equation on the spring, that causes the force to explode. I will double check everything on my end, I think I might have madea  mistake somehwere.