Fixed explosion problem.
The issue was that I did not set up the simulation using Helmut's vector calculus. Helmut did it using the sensible and mathematically orthodox way using vectors, whereas I used a more hacky bandaid solution. I think it won't make a difference to use Helmut's way in terms of speed or memory, so I will continue to use this way.