Radial Spider web reservoir computer working
The images are attached as PDF files. I am having an issue with Matlab's dynamic plotting for animation and some values of the input force break the system (which is what is expected I think) so I need some feedback and guidance, but I think this is a working spider web reservoir computer. Output signal = sine wave.