mx and mat working
Caveat: This is working only for Windows and configured for win64 and Matlab2017. The makefile is written for that and the .dll files (which are not being pushed) are all Win64 Matlab .dlls.
The makefile can be rewritten for Linux but it is an involved process. I can write a general how to and link to an example of a makefile for Linux, but aside from that I think this .mat import and export is just for my benefit. https://stackoverflow.com/questions/44564442/link-mat-h-in-a-c-file